added workshp form hook and split up components
